rafactor code to use SOLID principles
