+++ person.rb
@@ -1,10 +1,5 @@
 require './rental'
-
-class Nameable
-  def correct_name
-    raise NotImplementedError
-  end
-end
+require './nameable'
 
 class Person < Nameable
   attr_reader :id
@@ -37,32 +32,5 @@ class Person < Nameable
 
   def add_rentals(date, _person, book)
     Rental.new(date, self, book)
-  end
-end
-
-class Decorator < Nameable
-  def initialize(nameable)
-    @nameable = nameable
-    super()
-  end
-
-  def correct_name
-    @nameable.correct_name
-  end
-end
-
-class CapitalizeDecorator < Decorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
-
-class TrimmerDecorator < Decorator
-  def correct_name
-    if @nameable.correct_name.length > 10
-      @nameable.correct_name[0..9]
-    else
-      @nameable.correct_name
-    end
   end
 end
